+++ front/lib/api/poke/plugins/mcp_server_views/get-mcp-server-view-credentials.ts
