+++ front/lib/api/poke/plugins/mcp_server_views/index.ts
@@ -1,1 +1,1 @@
-export * from "./get-mcp-server-view-access-token";
+export * from "./get-mcp-server-view-credentials";
